requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
ccxt
plotly
tensorflow
# statsmodels

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
google-cloud-bigquery
google-auth
prophet
mlflow
s3fs
fastapi

# utilities
six>=1.14
joblib
memoized-property
termcolor
db-dtypes
